# openapi.yaml
openapi: 3.0.3
info:
  title: Racetag Backend API
  version: 0.1.0
servers:
  - url: http://localhost:8000
paths:
  /events/tag/batch:
    post:
      summary: Ingest a batch of tag events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagEventBatchDTO'
      responses:
        '200':
          description: Batch processed summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchIngestResultDTO'
...
components:
  schemas:
    TagEventDTO:
      type: object
      required: [source, reader_ip, reader_serial, timestamp, event_type, tag_id]
      properties:
        source:
          type: string
          example: sirit-510
        reader_ip:
          type: string
        reader_serial:
          type: string
        timestamp:
          type: string
          description: ISO8601 with milliseconds and Z
          example: 2025-10-16T11:53:32.298Z
        event_type:
          type: string
          enum: [arrive, depart]
        tag_id:
          type: string
          description: Uppercase hex without 0x
        session_id:
          type: integer
          nullable: true
        antenna:
          type: integer
          nullable: true
        rssi:
          type: integer
          nullable: true
        first:
          type: string
          nullable: true
        last:
          type: string
          nullable: true
    TagEventBatchDTO:
      type: object
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/TagEventDTO'
    BatchIngestResultDTO:
      type: object
      required: [events_processed]
      properties:
        events_processed:
          type: integer
...